Sensori completati
